Automatically deploy to server with fabric and Travis CI
This commit adds [fabric](fabfile.org) to automatically deploy to the
new virtual machine that wo-ist-markt runs on. It also changes the
Travis CI configuration to automatically run the deployment whenever a
push to master is triggered.

A slight change was needed to the `citylist-spec` because it was using
ecmascript 6 features that were not available on Travis CI yet.

More information can be found in the original pull request at
https://github.com/wo-ist-markt/wo-ist-markt.github.io/pull/115