Display markets using aggregated JSON files.
Always display markets from all cities that match the current selection
(now/today/other) using the JSON files produced by `npm run build`. The
layer selection controls are dynamically updated whenever the view
changes (i.e. when the map is moved or zoomed): The controls are
enabled/disabled based on which markers are available in the currently
displayed part of the map (instead of the currently selected city) and
the setting is automatically adjusted if the current view doesn't
contain any markets in the selected category.