Fixing and improving test of city objects in city list
As the city list is an object, cities.length does not work.  Therefore
now Object.keys() is used to get only the keys and iterate over them,
similar to other tests.

If a property for one city is missing, a error message is used to help
understanding the error.

Additional a test has been added for each city object to make sure that
it has not more properties than the required two.