Configure netlify to build project
By adding the `netlify.toml` file, we can tell netlify which
command to run in order to create the javascript bundle.
We also specify which node verison we want to use (needs to
be compatible with the rollup script) and which Python version
to use. Note that we don't technically need Python but netlify
notices that there is a `requirements.txt` and therefore tries
to install the requirements. Since those requirements aren't
available for the netlify default Python version 2.7, the build
would fail at that step. By specifing a more recent version which
matches the requirements, the install completes fine (although
it does not actually do anything with those requirements after
installing...).